+++ c4actor-extra/src/main/scala/ee/cone/c4actor/CommonRequestUtility.scala
@@ -1,0 +1,5 @@
+package ee.cone.c4actor
+
+trait CommonRequestUtility {
+
+}
