+++ c4actor-extra/src/main/scala/ee/cone/c4actor/CtxType.scala
@@ -1,0 +1,8 @@
+package ee.cone.c4actor
+
+
+object CtxType {
+  type Ctx = Map[Request, _]
+  type Request = Product
+  type ContextId = String
+}
