+++ c4actor-extra/src/main/scala/ee/cone/c4actor/DepAssembleUtility.scala
@@ -1,0 +1,28 @@
+package ee.cone.c4actor
+
+import java.nio.ByteBuffer
+import java.nio.charset.StandardCharsets
+import java.util.UUID
+
+import ee.cone.c4actor.CtxType.Request
+import ee.cone.c4actor.Types.SrcId
+
+trait DepAssembleUtility {
+  def generatePK(rq: Request, adapterRegistry: QAdapterRegistry): SrcId
+
+  def stringToKey(value: String): SrcId
+}
+
+trait DepAssembleUtilityImpl extends DepAssembleUtility {
+  def generatePK(rq: Request, adapterRegistry: QAdapterRegistry): SrcId = {
+    val valueAdapter = adapterRegistry.byName(rq.getClass.getName)
+    val bytes = valueAdapter.encode(rq)
+    UUID.nameUUIDFromBytes(toBytes(valueAdapter.id) ++ bytes).toString
+  }
+
+  private def toBytes(value: Long) =
+    ByteBuffer.allocate(java.lang.Long.BYTES).putLong(value).array()
+
+  def stringToKey(value: String) =
+    UUID.nameUUIDFromBytes(value.getBytes(StandardCharsets.UTF_8)).toString
+}
