+++ c4actor-extra/src/main/scala/ee/cone/c4actor/Resolvable.scala
@@ -1,0 +1,7 @@
+package ee.cone.c4actor
+
+import ee.cone.c4actor.CtxType.Request
+
+case class Resolvable[+A](value: Option[A], requests: Seq[Request] = Nil)
+
+case class UpResolvable(request: RequestWithSrcId, resolvable: Resolvable[_])
