+++ c4actor-extra/src/main/scala/ee/cone/c4actor/Response.scala
@@ -1,0 +1,5 @@
+package ee.cone.c4actor
+
+import ee.cone.c4actor.Types.SrcId
+
+case class Response(request: RequestWithSrcId, value: Option[_], rqList: List[SrcId] = Nil)
