--- c4actor-extra/src/main/scala/ee/cone/c4actor/dependancy/FilterRequest.scala
@@ -1,19 +1,0 @@
-package ee.cone.c4actor.dependancy
-
-import ee.cone.c4proto.{Protocol, protocol}
-
-class FilterRequest {
-
-}
-
-@protocol object FilterRequestProtocol extends Protocol {
-
-
-}
-
-//@Id(0x0fa0)
-//@Id(0x0fa1)
-//@Id(0x0fa2)
-//@Id(0x0fa3)
-//@Id(0x0fa4)
-//@Id(0x0fa5)
