--- c4actor-extra/src/main/scala/ee/cone/c4actor/dependancy/Response.scala
@@ -1,5 +1,0 @@
-package ee.cone.c4actor.dependancy
-
-import ee.cone.c4actor.Types.SrcId
-
-case class Response(request: RequestWithSrcId, value: Option[_], rqList: List[SrcId] = Nil)
