--- c4actor-extra/src/main/scala/ee/cone/c4actor/dependancy/package.scala
@@ -1,8 +1,0 @@
-package ee.cone.c4actor
-
-
-package object CtxType {
-  type Ctx = Map[Request, _]
-  type Request = Product
-  type ContextId = String
-}
